backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.5.3
Flask-Limiter==2.8.1
werkzeug==3.0.1
supabase==2.3.0
python-dotenv==1.0.0

